# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ApiGatawey",
        "title": "ApiGatawey",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:55555",
    "basePath": "/",
    "paths": {
...
        "/lifestyle/create": {
            "post": {
                "description": "Create a new Lifestyle record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lifestyle"
                ],
                "summary": "CreateLifeStyle",
                "parameters": [
                    {
                        "description": "Create Lifestyle",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.LifestyleCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/health_sync.Void"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lifestyle/delete/{id}": {
            "delete": {
                "description": "Delete a Lifestyle record based on the provided id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lifestyle"
                ],
                "summary": "DeleteLifeStyle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lifestyle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.GetById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lifestyle/get/{id}": {
            "get": {
                "description": "Retrieve a Lifestyle record based on the provided id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lifestyle"
                ],
                "summary": "GetLifeStyle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.GetById"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lifestyle/list": {
            "get": {
                "description": "List Lifestyle records with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lifestyle"
                ],
                "summary": "ListLifeStyle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.LifestyleList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/lifestyle/update/{id}": {
            "put": {
                "description": "Update a Lifestyle record based on the provided id and user_id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lifestyle"
                ],
                "summary": "UpdateLifeStyle",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lifestyle Update Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.LifestyleUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.LifestyleUpdate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "health_sync.GetById": {
            "type": "object",
            "properties": {
                "Id": {
                    "type": "string"
                }
            }
        },
...
        "health_sync.Lifestyle": {
            "type": "object",
            "properties": {
                "date_type": {
                    "type": "string"
                },
                "date_value": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recorded_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.LifestyleCreate": {
            "type": "object",
            "properties": {
                "date_type": {
                    "type": "string"
                },
                "date_value": {
                    "type": "string"
                },
                "recorded_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.LifestyleList": {
            "type": "object",
            "properties": {
                "LifestyleList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/health_sync.Lifestyle"
                    }
                }
            }
        },
        "health_sync.LifestyleUpdate": {
            "type": "object",
            "properties": {
                "date_type": {
                    "type": "string"
                },
                "date_value": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recorded_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "health_sync.Void": {
            "type": "object"
        },
...
    }
}